(PC-34600) feat(ui): move LoadingPage in pages folder
