+++ src/features/bookings/pages/BookingDetails/BookingDetails.tsx
@@ -39,7 +39,6 @@ import { useOpacityTransition } from 'ui/animations/helpers/useOpacityTransition
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
 import { ButtonTertiaryBlack } from 'ui/components/buttons/ButtonTertiaryBlack'
 import { HeaderWithImage } from 'ui/components/headers/HeaderWithImage'
-import { LoadingPage } from 'ui/components/LoadingPage'
 import { useModal } from 'ui/components/modals/useModal'
 import { SectionWithDivider } from 'ui/components/SectionWithDivider'
 import { Separator } from 'ui/components/Separator'
@@ -47,6 +46,7 @@ import { SNACK_BAR_TIME_OUT, useSnackBarContext } from 'ui/components/snackBar/S
 import { ExternalTouchableLink } from 'ui/components/touchableLink/ExternalTouchableLink'
 import { InternalTouchableLink } from 'ui/components/touchableLink/InternalTouchableLink'
 import { ViewGap } from 'ui/components/ViewGap/ViewGap'
+import { LoadingPage } from 'ui/pages/LoadingPage'
 import { EmailFilled } from 'ui/svg/icons/EmailFilled'
 import { getSpacing, TypoDS } from 'ui/theme'
 import { getHeadingAttrs } from 'ui/theme/typographyAttrs/getHeadingAttrs'
