+++ src/features/identityCheck/pages/identification/ubble/UbbleWebview.tsx
@@ -10,7 +10,7 @@ import { parseUrlParams } from 'features/identityCheck/pages/helpers/parseUrlPar
 import { navigateToHome } from 'features/navigation/helpers/navigateToHome'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
 import { analytics } from 'libs/analytics/provider'
-import { LoadingPage } from 'ui/components/LoadingPage'
+import { LoadingPage } from 'ui/pages/LoadingPage'
 import { Spacer } from 'ui/theme'
 
 // To avoid [Error: Unable to open URL: about:srcdoc. Add about to LSApplicationQueriesSchemes in your Info.plist.]
