+++ src/features/identityCheck/pages/identification/ubble/UbbleWebview.web.tsx
@@ -8,7 +8,7 @@ import { navigateToHome } from 'features/navigation/helpers/navigateToHome'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
 import { analytics } from 'libs/analytics/provider'
 import { Helmet } from 'libs/react-helmet/Helmet'
-import { LoadingPage } from 'ui/components/LoadingPage'
+import { LoadingPage } from 'ui/pages/LoadingPage'
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 declare const Ubble: any
